alter models and update frontend

+++ backend/aqms/models.py
@@ -9,15 +9,15 @@ class PowerManager(models.Manager):
 
 
 class Aqms(models.Model):
-    ppm = models.FloatField()
+    ppm = models.DecimalField(max_digits=6, decimal_places=3)
 
-    vol_pz = models.FloatField()
-    amp_pz = models.FloatField()
-    wat_pz = models.FloatField()
+    vol_pz = models.DecimalField(max_digits=8, decimal_places=6)
+    amp_pz = models.DecimalField(max_digits=8, decimal_places=6)
+    wat_pz = models.DecimalField(max_digits=8, decimal_places=6)
 
-    vol_wt = models.FloatField()
-    amp_wt = models.FloatField()
-    wat_wt = models.FloatField()
+    vol_wt = models.DecimalField(max_digits=6, decimal_places=3)
+    amp_wt = models.DecimalField(max_digits=6, decimal_places=3)
+    wat_wt = models.DecimalField(max_digits=6, decimal_places=3)
 
     windspeed = models.SmallIntegerField()
     date_time = models.DateTimeField(auto_now_add=True)
